requirements/dist.txt
setuptools>=20.6.7
wheel>=0.29.0
flake8>=2.5.4
flake8-bugbear
flake8-blind-except
flake8-comprehensions
flake8-pyi
flake8-import-order
tox>=2.3.1
sphinx2rst>=1.0
bumpversion
twine
pytest-sugar
